Se Termino la vista de DetallerRutina

+++ src/main/java/org/dev_busters/FitBody_Manager/FitBody/Manager/persistence/entity/EjercicioEntity.java
@@ -17,7 +17,7 @@ import org.dev_busters.FitBody_Manager.FitBody.Manager.dominio.enums.tipoEjercic
 public class EjercicioEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idEjercicio;
+    private Long idEjercicio;
 
     @Column(length = 64, nullable = false)
     private String nombreEjercicio;
@@ -34,11 +34,11 @@ public class EjercicioEntity {
     @Enumerated(EnumType.STRING)
     private Dificultad dificultad;
 
-    public int getIdEjercicio() {
+    public Long getIdEjercicio() {
         return idEjercicio;
     }
 
-    public void setIdEjercicio(int idEjercicio) {
+    public void setIdEjercicio(Long idEjercicio) {
         this.idEjercicio = idEjercicio;
     }
